+++ pkg/dashboard/apiserver/common/common.go
@@ -231,24 +231,18 @@ func (s *Service) listNamespaces(c *gin.Context) {
 // @Router /common/chaos-available-namespaces [get]
 // @Failure 500 {object} utils.APIError
 func (s *Service) getChaosAvailableNamespaces(c *gin.Context) {
-	kubeCli, err := clientpool.ExtractTokenAndGetClient(c.Request.Header)
-	if err != nil {
-		_ = c.Error(utils.ErrBadRequest.WrapWithNoMessage(err))
-		return
-	}
-
 	var namespaces sort.StringSlice
 
 	if s.conf.ClusterScoped {
 		var nsList v1.NamespaceList
-		if err := kubeCli.List(context.Background(), &nsList); err != nil {
+		if err := s.kubeCli.List(context.Background(), &nsList); err != nil {
 			c.Status(http.StatusInternalServerError)
 			_ = c.Error(utils.ErrInternalServer.WrapWithNoMessage(err))
 			return
 		}
 		namespaces = make(sort.StringSlice, 0, len(nsList.Items))
 		for _, ns := range nsList.Items {
-			if s.conf.EnableFilterNamespace && !namespace.CheckNamespace(context.TODO(), kubeCli, ns.Name) {
+			if s.conf.EnableFilterNamespace && !namespace.CheckNamespace(context.TODO(), s.kubeCli, ns.Name, u.Log) {
 				continue
 			}
 			namespaces = append(namespaces, ns.Name)